requirements.txt
Flask==1.1.2
line-bot-sdk==1.7.2
requests==2.25.1